@@ -58,9 +58,9 @@
     "@types/classnames": "^2.3.1",
     "@types/deep-equal": "^1.0.1",
     "@types/express": "^4.17.17",
-    "@types/react": "^17.0.62",
+    "@types/react": "^18.2.48",
     "@types/react-collapse": "^5.0.1",
-    "@types/react-dom": "^17.0.20",
+    "@types/react-dom": "^18.2.18",
     "@types/react-modal": "^3.16.0",
     "@types/react-router-dom": "^5.3.3",
     "@types/react-router-hash-link": "^2.4.6",
@@ -83,9 +83,9 @@
     "nav-frontend-core": "^6.0.1",
     "nav-hoykontrast": "^2.1.2",
     "query-string": "^8.1.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-collapse": "^5.1.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-modal": "^3.3.1",
     "react-router-dom": "^6.21.0",
     "react-router-hash-link": "^2.4.3",
