@@ -1,4 +1,4 @@
-import type { ReactNode, JSX } from 'react';
+import type { ReactNode, JSX, PropsWithChildren } from 'react';
 import React, { useEffect, useState } from 'react';
 
 import type { AxiosRequestConfig } from 'axios';
@@ -39,7 +39,7 @@ const initalState: IModal = {
     visModal: false,
 };
 
-interface IProps {
+interface IProps extends PropsWithChildren {
     autentisertSaksbehandler: ISaksbehandler | undefined;
 }
 
@@ -243,7 +243,7 @@ const [AppContentProvider, useApp] = createUseContext(() => {
     };
 });
 
-const AuthOgHttpProvider: React.FC = ({ children }) => {
+const AuthOgHttpProvider: React.FC<PropsWithChildren> = ({ children }) => {
     const { innloggetSaksbehandler, settAutentisert } = useAuth();
 
     return (
