@@ -1,3 +1,4 @@
+import type { PropsWithChildren } from 'react';
 import React from 'react';
 
 import styled from 'styled-components';
@@ -8,7 +9,7 @@ const Container = styled.div`
     margin-top: 1rem;
 `;
 
-const Knapperekke: React.FC = ({ children }) => {
+const Knapperekke: React.FC<PropsWithChildren> = ({ children }) => {
     return <Container>{children}</Container>;
 };
 
