@@ -1,3 +1,4 @@
+import type { PropsWithChildren } from 'react';
 import React from 'react';
 
 import styled from 'styled-components';
@@ -7,6 +8,8 @@ const StyledLegend = styled.legend`
     clip: rect(0 0 0 0);
 `;
 
-const SkjultLegend: React.FC = ({ children }) => <StyledLegend>{children}</StyledLegend>;
+const SkjultLegend: React.FC<PropsWithChildren> = ({ children }) => (
+    <StyledLegend>{children}</StyledLegend>
+);
 
 export default SkjultLegend;
